IOS-6969 Move tests
